NEW/requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
pyod==1.1.3
scapy==2.5.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1
networkx==3.2.1
paho-mqtt==1.6.1
pyyaml==6.0.1
sqlalchemy==2.0.25
psutil==5.9.8
requests==2.31.0
python-dotenv==1.0.1
